+++ CHANGES.rst
@@ -30,6 +30,8 @@
 
 - Added misc. required db keyword attributes. [JIRA RAD-7]
 
+- Added wfi photom schema and tests. [#34]
+  
 - Added Dark schema and updated Flat schema. [#35]
 
 - Added dq schema. [#32]
@@ -41,7 +43,6 @@
 - Updated aperture, basic, ephemeris, exposure, guidestar, observation, pixelarea, and visit schemas. [#46]  
 
 - Added support for variance object schemas. [#38] 
-
 
 0.1.0 (unreleased)
 ==================